The SDK classes are split in their own files to follow best practices an allow autoloading.
