+++ src/GSException.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Softonic\Gigya;
+
+use Exception;
+
+/**
+ * Gigya Socialize Exception
+ */
+class GSException extends Exception
+{
+    public $errorMessage;
+}
